+++ rbx_reflector/src/defaults.rs
@@ -37,7 +37,7 @@ pub fn apply_defaults(
             continue;
         }
 
-        found_classes.insert(instance.class.clone());
+        found_classes.insert(instance.class);
 
         apply_instance_defaults(database, instance);
     }
@@ -59,7 +59,7 @@ fn apply_instance_defaults(database: &mut ReflectionDatabase, instance: &Instanc
     };
 
     for (property_name, property_value) in &instance.properties {
-        let property_name = Cow::Owned(property_name.clone());
+        let property_name = Cow::Owned(property_name.to_string());
 
         match property_value.ty() {
             // We skip the Ref type because its default value is not useful.