+++ rbx_xml/CHANGELOG.md
@@ -1,8 +1,10 @@
 # rbx_xml Changelog
 
 ## Unreleased
+* Improved performance of serializer and deserializer by using `Ustr` to represent property and class names ([#462]).
 * `Content` data now serializes with `ContentId`, reflecting Roblox's changes. ([#455])
 
+[#462]: https://github.com/rojo-rbx/rbx-dom/pull/462
 [#455]: https://github.com/rojo-rbx/rbx-dom/pull/455
 
 ## 0.13.5 (2024-08-22)